tests/wpt/tests/encoding-detection/ko-EUC-KR-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>ko EUC-KR late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "EUC-KR", 'Expected EUC-KR');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/ko-EUC-KR-late.sub.html");
</script>
